+++ odpi-simple/ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for odpi-simple
 
+## 0.4.0.0  -- 2021-10-24
+
+* Replace `Proxy` in typeclasses with AlowAmbiguousTypes and TypeApplications
+* Generic FromRow for simple products
+
 ## 0.3.0.0  -- 2021-05-08
 
 * Update to odpi-libdpi 0.3.